:recycle: refactor(admin): change apis dir to common, domains
