--- backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/apis/common/HealthController.java
@@ -1,12 +1,0 @@
-package com.depromeet.apis.common;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HealthController {
-    @GetMapping({"/health", "/"})
-    public String health() {
-        return "OK";
-    }
-}
