--- backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/apis/user/controller/UserController.java
@@ -1,41 +1,0 @@
-package com.depromeet.apis.user.controller;
-
-
-import com.depromeet.apis.user.dto.ResponseDto;
-import com.depromeet.apis.user.dto.UserAllResponseDto;
-import com.depromeet.apis.user.dto.UserDetailResponseDto;
-import com.depromeet.apis.user.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.web.PageableDefault;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/users")
-@RequiredArgsConstructor
-public class UserController {
-    private final UserService userService;
-
-    @GetMapping
-    public ResponseEntity<UserAllResponseDto> getAllUsers(
-            @PageableDefault(size = 20, page = 0, sort = "createdAt",
-                    direction = Sort.Direction.DESC) Pageable pageable
-    ) {
-        var response = userService.getAllUsers(pageable);
-        return ResponseDto.ok(response);
-    }
-
-
-    @GetMapping("/{userId}")
-    public ResponseEntity<UserDetailResponseDto> getUserDetail(
-            @PathVariable("userId") Long userId
-    ) {
-        var response = userService.getUserDetail(userId);
-        return ResponseDto.ok(response);
-    }
-}
