--- backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/apis/user/dto/UserActivityResponseDto.java
@@ -1,13 +1,0 @@
-package com.depromeet.apis.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class UserActivityResponseDto {
-    private String title;
-    private String content;
-}
