--- backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/apis/user/dto/UserAllResponseDto.java
@@ -1,16 +1,0 @@
-package com.depromeet.apis.user.dto;
-
-import com.depromeet.apis.common.dto.PageMetaData;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class UserAllResponseDto {
-    List<UserResponseDto> users;
-    PageMetaData meta;
-}
