--- backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/apis/user/dto/UserAllStaticCountDto.java
@@ -1,14 +1,0 @@
-package com.depromeet.apis.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class UserAllStaticCountDto {
-    private Long allUserCount;
-    private int todayUserCount;
-    private int dropUserCount;
-}
