--- backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/apis/user/dto/UserDetailInfoResponseDto.java
@@ -1,15 +1,0 @@
-package com.depromeet.apis.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class UserDetailInfoResponseDto {
-    private int allDropCount;
-    private String mainDropLocation;
-    private String dropLocations;
-    private String interestGenre;
-}
