--- backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/apis/user/dto/UserDetailResponseDto.java
@@ -1,16 +1,0 @@
-package com.depromeet.apis.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class UserDetailResponseDto {
-    private UserBasicInfoResponseDto userBasicInfo;
-    private UserDetailInfoResponseDto userDetailInfo;
-    private List<UserActivityResponseDto> userActivity;
-}
