--- backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/apis/user/dto/UserSignUpCountRequestDto.java
@@ -1,19 +1,0 @@
-package com.depromeet.apis.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class UserSignUpCountRequestDto {
-    private LocalDateTime startDate = LocalDateTime.of(LocalDate.now().minusDays(7), LocalTime.MIN);
-    private LocalDateTime endDate = LocalDateTime.now();
-}
