--- backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/apis/user/dto/UserSignUpCountResponseDto.java
@@ -1,4 +1,0 @@
-package com.depromeet.apis.user.dto;
-
-public record UserSignUpCountResponseDto(String date, Long count) {
-}
