--- backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/apis/village/dto/ItemCountPerVillageResponseDto.java
@@ -1,11 +1,0 @@
-package com.depromeet.apis.village.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class ItemCountPerVillageResponseDto {
-    private String villageName;
-    private Long itemCount;
-}
