+++ backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/common/HealthController.java
@@ -1,0 +1,12 @@
+package com.depromeet.common;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HealthController {
+    @GetMapping({"/health", "/"})
+    public String health() {
+        return "OK";
+    }
+}
