+++ backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/common/dto/PageMetaData.java
@@ -1,0 +1,15 @@
+package com.depromeet.common.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class PageMetaData {
+    private int page;
+    private int size;
+    private int totalElements;
+    private int totalPages;
+}
