+++ backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/domains/user/dto/UserActivityResponseDto.java
@@ -1,0 +1,13 @@
+package com.depromeet.domains.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class UserActivityResponseDto {
+    private String title;
+    private String content;
+}
