+++ backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/domains/user/dto/UserDetailInfoResponseDto.java
@@ -1,0 +1,15 @@
+package com.depromeet.domains.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class UserDetailInfoResponseDto {
+    private int allDropCount;
+    private String mainDropLocation;
+    private String dropLocations;
+    private String interestGenre;
+}
