+++ backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/domains/user/dto/UserDetailResponseDto.java
@@ -1,0 +1,16 @@
+package com.depromeet.domains.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class UserDetailResponseDto {
+    private UserBasicInfoResponseDto userBasicInfo;
+    private UserDetailInfoResponseDto userDetailInfo;
+    private List<UserActivityResponseDto> userActivity;
+}
