+++ backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/domains/user/dto/UserSignUpCountResponseDto.java
@@ -1,0 +1,4 @@
+package com.depromeet.domains.user.dto;
+
+public record UserSignUpCountResponseDto(String date, Long count) {
+}
