+++ backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/domains/village/dto/ItemCountPerVillageResponseDto.java
@@ -1,0 +1,11 @@
+package com.depromeet.domains.village.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class ItemCountPerVillageResponseDto {
+    private String villageName;
+    private Long itemCount;
+}
